Bug #33544: UDF_INIT member decimals initialized wrong with   STRING_RESULT argument
There is a "magic" number for precision : NOT_FIXED_DEC. 
This means that the precision is not a fixed number.
But this constant was re-defined in several files and 
was not available to the UDF developers.

Moved the NOT_FIXED_DEC definition to the correct header
and removed the redundant definitions.