[patch 45/64] Always use TLS functions instead of __declspec(thread) to access thread local storage variables.(patch by wlad)
but the change removes the necessity to recomplile the same source
files twice -  with USE_TLS for DLLs and without USE_TLS for EXEs.
Real benefit of this change is better readability and maintainability
of TLS functions within MySQL.
      
There is a performance loss using TlsXXX functions compared to __declspec 
but the difference is negligible in practice. In a sysbench-like benchmark 
I ran with with TlsGetValue, pthread_[get|set]_specific was called 600000000 
times and took 0.17sec of total 35min CPU time, or 0.008%.
modified:
  client/CMakeLists.txt
  include/my_pthread.h
  libmysql/CMakeLists.txt
  libmysqld/CMakeLists.txt
  libmysqld/examples/CMakeLists.txt
  mysys/my_thr_init.c
  mysys/my_winthread.c
  tests/CMakeLists.txt

per-file messages:
  client/CMakeLists.txt
    Remove USE_TLS and mysqlclient_notls
  include/my_pthread.h
    Remove USE_TLS and mysqlclient_notls
  libmysql/CMakeLists.txt
    Remove USE_TLS and mysqclient_notls
  libmysqld/CMakeLists.txt
    Remove USE_TLS
  libmysqld/examples/CMakeLists.txt
    Remove USE_TLS
  mysys/my_thr_init.c
    Remove use of USE_TLS
  mysys/my_winthread.c
    Remove win_pthread_setspecific
  tests/CMakeLists.txt
    Remove mysqclient_notls
=== modified file 'client/CMakeLists.txt'