Bug# 39645 Losing size info for VARCHAR/CHAR for disk-data could segfault API apps
VAR* types are stored fixed-size on disk.  This var->fixed mapping occurs 
at the NDBAPI level, and some evidence of it can leak out to NDBAPI applications.

This patch modifies the implementation of a number of NDBAPI methods so that
disk-based VAR* types display the same behaviour as memory based VAR* types.

From mysql-5.1-telco-6.4, the handling of VAR* types on disk is changed by 
WL4499, and this patch becomes unnecessary.

HugoTest table D2 is augmented with some VARCHAR and LONGVARCHAR columns.