WL#4350 Reconfigure transporter
Make it possible to run IPCConfig::configureTransporters
also to reconfigure the transporters in TransporteRegistry.

If the transporter already exist, call "configure(conf)" on it.

Currently the Transporter only detect if the conf is same or
different. If different a warning about this will be printed
to log and the flag m_need_restart is set.

Also detect if a transporter should be removed and print a
warning about restart in same fashion as above.